Adding Navigation controls. Need to clean up (in this PR) and refine (in the next)
