+++ src/Header.jsx
@@ -1,13 +1,13 @@
 import React, { Component } from "react";
 import "./Header.css";
-import Navigation from "./Navigation";
+import NavBar from "./NavBar";
 
 class Header extends Component {
     render() {
         return (
             <div className='wrapper'>
-                <div className='header'> Cistern</div>
-                <Navigation></Navigation>
+                <div className='header'>Cistern</div>
+                <NavBar />
             </div>
         );
     }
