+++ src/NavBar.css
@@ -1,0 +1,55 @@
+.navBar-container {
+    display: inline-block;
+    padding-top: 20px;
+    margin-right: -200px;
+}
+.navBar-container > * {
+    padding-inline: 10px;
+    text-decoration: none;
+    font-variant: small-caps;
+    font-size: large;
+}
+.linkText {
+    color: black;
+}
+.linkTextClosed {
+    color: purple;
+}
+/* .sidenav {
+    width: 250px;
+    transition: width 0.3s ease-in-out;
+    height: 100vh;
+    background-color: rgb(10, 25, 41);
+    padding-top: 28px;
+}
+.sidenavClosed {
+    transition: width 0.3s ease-in-out;
+    width: 60px;
+}
+.sideitem {
+    display: flex;
+    align-items: center;
+    padding: 10px 20px;
+    cursor: pointer;
+    color: #b2bac2;
+    text-decoration: none;
+} */
+/* .linkText {
+    padding-left: 16px;
+}
+.linkTextClosed {
+    visibility: hidden;
+}
+.sideitem:hover {
+    background-color: #244f7d1c;
+}
+.menuBtn {
+    align-self: center;
+    align-self: flex-start;
+    justify-self: flex-end;
+    color: #b2bac2;
+    background-color: transparent;
+    border: none;
+    cursor: pointer;
+    padding-left: 20px;
+} */
