+++ src/Pages/ExploreStyles/Popup.css
