+++ src/Pages/People.jsx
@@ -1,0 +1,3 @@
+export default function People() {
+    return <div>Settings</div>;
+}
