+++ src/Pages/Statistics.jsx
@@ -1,0 +1,3 @@
+export default function Statistics() {
+    return <div>Statistics</div>;
+}
